refactor: remove unused exported functions, types, and files
