+++ slog/context.go
@@ -6,11 +6,6 @@ type contextKey string
 
 const loggerContextKey = contextKey("logger")
 
-// WithLogger returns a new context with a logger added to it.
-func WithLogger(ctx context.Context, l Logger) context.Context {
-	return context.WithValue(ctx, loggerContextKey, l)
-}
-
 // GetLogger returns a logger associated with this context.
 func GetLogger(ctx context.Context) Logger {
 	v := ctx.Value(loggerContextKey)
@@ -21,9 +16,4 @@ func GetLogger(ctx context.Context) Logger {
 	logger, _ := v.(Logger)
 
 	return logger
-}
-
-// With adds logging metadata to the logger within the context.
-func With(ctx context.Context, key string, value any) context.Context {
-	return WithLogger(ctx, GetLogger(ctx).With(key, value))
 }
