+++ util/gitutil/detectgit.go
@@ -433,19 +433,6 @@ func detectGitAuthor(ctx context.Context, dir string, format string) (string, er
 	return strings.SplitN(outStr, "\n", 2)[0], nil
 }
 
-// ConfigEmail returns the user's currently configured (global) email address.
-func ConfigEmail(ctx context.Context) (string, error) {
-	cmd := exec.CommandContext(ctx, "git", "config", "--get", "user.email")
-	cmd.Stderr = nil // force capture of stderr on errors
-
-	out, err := cmd.Output()
-	if err != nil {
-		return "", fmt.Errorf("detect git global email: %w", err)
-	}
-
-	return strings.TrimSpace(string(out)), nil
-}
-
 func detectGitCoAuthors(ctx context.Context, dir string) ([]string, error) {
 	cmd := exec.CommandContext(ctx, "git", "log", "-1", "--format=%b")
 	cmd.Dir = dir
