+++ util/platutil/native.go
@@ -7,24 +7,8 @@ import (
 
 	"github.com/containerd/platforms"
 	"github.com/moby/buildkit/client"
-	gwclient "github.com/moby/buildkit/frontend/gateway/client"
 	specs "github.com/opencontainers/image-spec/specs-go/v1"
 )
-
-// GetNativePlatform returns the native platform for a given gwClient.
-func GetNativePlatform(gwClient gwclient.Client) (specs.Platform, error) {
-	ws := gwClient.BuildOpts().Workers
-	if len(ws) == 0 {
-		return specs.Platform{}, errors.New("no worker found via gwclient")
-	}
-
-	nps := ws[0].Platforms
-	if len(nps) == 0 {
-		return specs.Platform{}, errors.New("no platform found for worker via gwclient")
-	}
-
-	return platforms.Normalize(nps[0]), nil
-}
 
 // GetNativePlatformViaBkClient returns the native platform for a given buildkit client.
 func GetNativePlatformViaBkClient(ctx context.Context, bkClient *client.Client) (specs.Platform, error) {
